Add temporary WR until PROD-13167 become fixes
Current installed verion of python-pymysql (0.7.2) has an issue with time,
as result event if time on nodes are ok, there can be situation taht
cinder-manage show that it is different. As result we've got the situation,
that volume become cleaned before it should be.
Upstream issue https://bugs.launchpad.net/cinder/+bug/1641312

So we need to install version 0.7.6 until it's fixed officially

Change-Id: I00cf611b7b07cf45f658c5a5eff60ab5eae169b7
Related: PROD-13167