Set grab_versions and snapshot fixtures as a dependency for salt_deployed
Sometimes, order of the fixtures is built in such way that
pytest tries to complete 'salt_deployed' before 'grab_versions'.
This lead to missing logs after the test is completed or failed.

Explicit set grab_versions to run before the salt_deployed, instead
of relying on floating order with 'autouse'.
The same for the fixture 'snapshot'

Change-Id: I4098413e0c6a016442630334c04bce218cb7f479