Set keepalive interval for paramiko SSHClient
If the ssh server has set the option ClientAliveInterval, it will
send an 'empty' packet to a client and closes the connection if
the client is not responding.
Looks like paramiko ignores such keepalive packages from server.

But paramiko has an option similar to ServerAliveInterval, which
allows to send an 'empty' package from the client to the server,
so the server will never reach it's own timeout
for ClientAliveInterval, and the connection won't be closed.

- export SSH_SERVER_ALIVE_INTERVAL=60 # 60 is default, 0 to disable
- remove workarounds for ClientAliveInterval. The same timeouts may
  be happened inside the platform and should be catched by the tests

Change-Id: I552cb7fdca59b51a3fc0e9e7d2cff7f28cb444dc