[mitaka-trusty] Restore ability to deploy mitaka release on trusty
As maintenance team we still responsible for one mitaka-trusty cloud.
In case of incoming bugfix-request we need an ability to deploy mitaka
cluster using trusty-based mitaka packages and environment.

Some times ago (last successful run was Oct.9) this deployment were broken.
This was caused by removal of trusty iscsi daemon's name from formulas
(this was expectable but still painful).

To ensure that we will not get deployment errors like this (which will be
because trusty deployment even not legacy) we want to bind our deployment jobs
to last successful deployment than we can get. After some investigation this combination of
packages/formulas/model/deployment steps was found.

We want to use 2018.8.0 MCP release with some hacks inside tcp-qa deployment process.
This SHOULD NOT affects any existing deployment but will give us a chance to verify
shipped packages in the future.

Change-Id: I0d067c2c85a61714621aab6233dcfe79d55c4a18