Lookup nodes and networks in the nested stacks
Each custom resource, defined in the "resource_registry:" key
in the heat environment file, creates a separated stack that
handles the resources defined in this custom resources file.
To get the nested stacks: openstack stack list --nested

To find all the OS::Nova::Server and OS::Neutron::Subnet,
need to inspect all the nested stack, check that the stacks
are childs from the current stack, and collect all the resources
from them.

Change-Id: I28f9dd16070104b8f0715987408b8c2c2f7d810e